Added gui_settings and widgets modules to allow for easier editing. Incorperated Graph class into taking over nodes and edge tracking for adding, deleting and adjaceny lists, including directed and undirected edges.
