+++ gui_settings.py
@@ -1,0 +1,22 @@
+from pygame import Color as color
+# Surface Dimensions
+SCREEN_WIDTH = 1600
+SCREEN_HEIGHT = 1024
+BUFFER_WIDTH = SCREEN_WIDTH
+BUFFER_HEIGHT = 64
+
+# Render Settings
+FPS = 60
+
+# Colors
+COLOR_NODE = color("White")
+COLOR_EDGE = color("aqua")
+COLOR_TEXT = color("black")
+COLOR_SELECTED = color("coral")
+COLOR_SCREEN = color("black")
+COLOR_WIDGET = color("SlateGray")
+
+# Mouse button variables
+LMB = 1
+MMB = 2
+RMB = 3
